+++ .github/workflows/python-app.yml
@@ -27,7 +27,6 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install black flake8
           pip install -r requirements.txt
           pip install -r requirements-dev.txt
       - name: Lint with flake8
@@ -50,7 +49,6 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pytest pytest-qt
         pip install -r requirements.txt
         pip install -r requirements-dev.txt
     - name: Test with pytest
@@ -70,7 +68,6 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pytest pytest-qt
         pip install -r requirements.txt
         pip install -r requirements-dev.txt
     - name: Test with pytest
@@ -101,7 +98,6 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pytest pytest-qt pytest-xvfb
         pip install -r requirements.txt
         pip install -r requirements-dev.txt
     - name: Test with pytest
