+++ requirements-dev-qt.txt
@@ -1,0 +1,2 @@
+pytest-qt
+pytest-xvfb
