+++ requirements-dev.txt
@@ -1,6 +1,4 @@
 pytest
 pytest-cov
-pytest-qt
-pytest-xvfb
 black
 flake8
