+++ src/core_atomistic_qt/qt_image.py
