+++ src/models/capedswcnt.py
@@ -5,8 +5,8 @@ import math
 import numpy as np
 from pathlib import Path
 
-from src_core_atomistic.atom import Atom
-from src_core_atomistic.atomic_model import AtomicModel as TAtomicModel
+from core_atomistic.atom import Atom
+from core_atomistic.atomic_model import AtomicModel as TAtomicModel
 from models.swnt import SWNT
