+++ src/models/swgnt.py
@@ -2,8 +2,8 @@
 
 import math
 
-from src_core_atomistic.atom import Atom
-from src_core_atomistic.atomic_model import AtomicModel
+from core_atomistic.atom import Atom
+from core_atomistic.atomic_model import AtomicModel
 
 
 class SWGNT(AtomicModel):
