+++ src/models/xsf.py
@@ -4,9 +4,9 @@ from copy import deepcopy
 import os
 import numpy as np
 
-from src_core_atomistic.atomic_model import AtomicModel as TAtomicModel
-from src_core_atomistic.periodic_table import TPeriodTable
-from src_core_atomistic import helpers
+from core_atomistic.atomic_model import AtomicModel as TAtomicModel
+from core_atomistic.periodic_table import TPeriodTable
+from core_atomistic import helpers
 
 from models.volumericdatablock import VolumericDataBlock
 from models.volumericdata import VolumericData
