+++ src/program/crystal.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from copy import deepcopy
 import numpy as np
-from src_core_atomistic.atom import Atom
-from src_core_atomistic.atomic_model import AtomicModel
-from src_core_atomistic import helpers
+from core_atomistic.atom import Atom
+from core_atomistic.atomic_model import AtomicModel
+from core_atomistic import helpers
 
 
 def model_0d_to_d12(model):
