+++ src/program/dftb.py
@@ -4,9 +4,9 @@ from copy import deepcopy
 
 import numpy as np
 
-from src_core_atomistic.atomic_model import AtomicModel
-from src_core_atomistic.periodic_table import TPeriodTable
-from src_core_atomistic import helpers
+from core_atomistic.atomic_model import AtomicModel
+from core_atomistic.periodic_table import TPeriodTable
+from core_atomistic import helpers
 
 
 def model_to_dftb_d0(model: AtomicModel):
