+++ src/program/qe.py
@@ -2,9 +2,9 @@
 import math
 from copy import deepcopy
 import numpy as np
-from src_core_atomistic import helpers
-from src_core_atomistic.atom import Atom
-from src_core_atomistic.atomic_model import AtomicModel
+from core_atomistic import helpers
+from core_atomistic.atom import Atom
+from core_atomistic.atomic_model import AtomicModel
 
 
 def vectors_from_pwout(f_name):
