+++ src/program/siesta.py
@@ -4,10 +4,10 @@ import xml.etree.ElementTree as ET
 
 import numpy as np
 
-from src_core_atomistic.atom import Atom
-from src_core_atomistic.atomic_model import AtomicModel
-from src_core_atomistic.periodic_table import TPeriodTable
-from src_core_atomistic import helpers
+from core_atomistic.atom import Atom
+from core_atomistic.atomic_model import AtomicModel
+from core_atomistic.periodic_table import TPeriodTable
+from core_atomistic import helpers
 
 
 class TSIESTA:
