+++ tests/gui4dft_opengl/test_guiopengl.py
@@ -1,4 +1,4 @@
-from src_core_atomistic.atomic_model import AtomicModel
+from core_atomistic.atomic_model import AtomicModel
 
 
 def test_guiopengl(guiopengl_widget):
