+++ tests_core_atomistic/conftest.py
@@ -1,5 +1,18 @@
-from src_core_atomistic.atomic_model import AtomicModel
+from pathlib import Path
+from core_atomistic.atomic_model import AtomicModel
+from core_atomistic.periodic_table import TPeriodTable
+
 import pytest
+
+
+@pytest.fixture
+def tests_path() -> Path:
+    return Path(__file__).parent
+
+
+@pytest.fixture
+def period_table() -> TPeriodTable:
+    return TPeriodTable()
 
 
 @pytest.fixture
