+++ tests_core_atomistic/test_atom.py
@@ -1,4 +1,4 @@
-from src_core_atomistic.atom import Atom
+from core_atomistic.atom import Atom
 
 
 def test_atom():
