+++ tests_core_atomistic/test_periodic_table.py
