+++ tests_core_atomistic_qt/test_pygrimagewidget.py
